Use native embedder for inference
